[step2] TicketOffice 클래스의 당첨 관련 기능을 LottoCompany 클래스로 분리
기존 TicketOffice의 역할이 과도하게 크다고 판단하여 당첨 번호 설정 및 당첨금 확인(통계) 기능을 담당하는 LottoCompany 클래스를 만들어 분리.

getStatistic()과 연결되는 calculateProfit에서 사용하는 TOTAL_PRICE 값은 TicketOffice에서 로또를 구매할 때 설정되기 때문에 구조 수정 필요.
클래스 분리로 인해 Main 로직도 수정 필요.
