+++ src/main/java/domain/TicketOffice.java
@@ -11,16 +11,6 @@ public class TicketOffice {
     private final int SELECTED_NUMBER = 6;
     private final int PRICE = 1000;
     private int TOTAL_PRICE;
-    private int bonusNumber;
-    private LottoTicket winningTicket;
-    private Map<Integer, Integer> statistics = new HashMap<>();
-    {
-        statistics.put(3, 0);
-        statistics.put(4, 0);
-        statistics.put(5, 0);
-        statistics.put(6, 0);
-        statistics.put(7, 0);
-    }
 
 
     public TicketOffice() {
@@ -56,46 +46,5 @@ public class TicketOffice {
             tickets.add(makeLottoTicket(isAuto));
         OutputView.completePurchase(numberOfTickets, change, tickets);
         return tickets;
-    }
-
-    public void setWinningTicket() {
-        winningTicket = new LottoTicket(InputView.getWinningNumber());
-    }
-
-    public void setBonusNumber() { // 추후 당첨 번호에 없는 번호만 받도록 예외처리 필요
-        bonusNumber = InputView.getBonusNumber();
-    }
-
-    public void getStatistic(List<LottoTicket> tickets) {
-        int matchedNumber = 0;
-        boolean isBonus = false;
-        for (LottoTicket ticket : tickets) {
-            matchedNumber = ticket.comparisonWinningTicket(winningTicket);
-            isBonus = ticket.checkBonusNumber(bonusNumber);
-            matchedNumber += temp(matchedNumber, isBonus);
-            statistics.computeIfPresent(matchedNumber, (k, v) -> v + 1);
-        }
-        OutputView.showWinningResult(this.statistics, calculateProfit(isBonus));
-    }
-
-    private int temp(int matchedNumber, boolean isBonus){
-        if (matchedNumber == 5){
-            if(isBonus){
-                return 2;
-            }
-        }
-        return 0;
-    }
-
-    public double calculateProfit(boolean isBonus) {
-        int totalPrize = 0;
-        for (Integer matchedNumber : statistics.keySet()) {
-            totalPrize += (switchPrize(matchedNumber, isBonus) * statistics.get(matchedNumber));
-        }
-        return ((double) (totalPrize - TOTAL_PRICE) / TOTAL_PRICE) * 100;
-    }
-
-    private int switchPrize(int matchedNumber, boolean isBonus) {
-        return Rank.designateRank(matchedNumber, isBonus).getPrize();
     }
 }
